Added error handling to Json and Jsonp serializers
Serializers must have error handling. When deserializing content the serializer might fail and the client needs to get a proper error message so that they can modify the content and try again. If serialization for some reason fails the client needs to get a proper serialized error message so that the client can handle the response without failing on their end.